+++ tests/fixtures/feature_materialize/initialize_new_columns_existing_table.sql
@@ -63,7 +63,7 @@ WITH ONLINE_REQUEST_TABLE AS (
 )
 SELECT
   AGG."cust_id",
-  CAST("_fb_internal_cust_id_window_w1800_sum_e8c51d7d1ec78e1f35195fc0cf61221b3f830295" AS FLOAT) AS "sum_30m_V220101"
+  CAST("_fb_internal_cust_id_window_w1800_sum_e8c51d7d1ec78e1f35195fc0cf61221b3f830295" AS DOUBLE) AS "sum_30m_V220101"
 FROM _FB_AGGREGATED AS AGG;
 
 SELECT
